restructured the project to simplify the hierarchy
