+++ kraken/spot/user.py
@@ -8,7 +8,7 @@
 from decimal import Decimal
 from typing import List, Optional, Union
 
-from ...base_api import KrakenBaseSpotAPI, defined, ensure_string
+from ..base_api import KrakenBaseSpotAPI, defined, ensure_string
 
 
 class User(KrakenBaseSpotAPI):
